scripts/requirements.txt
pyyaml==6.0
ruamel.yaml==0.18.0